JBTM-845 clean up recovery tests
